Added a better separation between synced and local created/modified files. Now local created/modified files survive re-mounts correctly!
Fixed file size report in the cache file system. Now files report the correct size in the remote filesystem, even
tough they are 0 sized in the local filesystem. The size of the remote file is stored in the log file for every
skeleton file, so we can access it fast and report it correctly inside readdir()!

Now the cache file system reports the correct size, uid, gid and time (access and modified) of the files in the remote
filesystem.

Properly implemented rmdir, so now when we remove a folder, it actually just flags it as removed, instead of removing it
for good. If the folder is re-created, only the flag is removed and the old folder shows up again!