It seems to be finally working reliably without deadlocks.
It's possible now to WRITE and OVERWRITE files to the filesystem. HradecFS will locally create the files in the local cache, as a normal local filesystem running on top of the remote filesystem. The local files survive re-mounts of the
filesystem (although it hasn't being throughly tested, for example, when local files are in folders that have been removed if the remote file system). Local files always have priority over remote files, even when remote files have being
modified after the local file. Since local files survive re-mounts, and HradecFS only updates the local cache at mount time, local files will NEVER reflect the data from updated remote files of the same name. The only way now to remove a
local file so the remote content of that file shows up, is removing the file and re-mounting. (since HradecFS re-checks for not-existent files at mount, it will retrieve the remote file properly). This local-file mechanism needs to be
better handled in the future, but that will only happen with the multi-threaded cache update mechanism!

Fixed absolute symlinks to symply fix the linked path by prefixing the mount folder path to the link. (a more inteligent handling is planned, with command link options to setup how to handle it)

Fixed loads of deadlocks in very special and unusual conditions.

Temporarely fixed long file name on control files by doing a string.right(210) of the control file name. We need to work on a better control file mechanism. (maybe a sqlite db, converting file names to a hash number or a multi-folder
mechanism)

Added a way to selectively turn on/off different levels of debugging through /tmp/.bbfs.log pipe file. (look at log.c for the bit-wise selection - for example, echo 8 > /tmp/.bbfs.log will turn off all log but file copy from remote and
local file creation)

Initiated code cleanup, preparing for the cache update mechanism. (no update of the cache yet, unless when re-mounting the filesystem)

The only way to trigger an update of a single file or folder now is:

1. file - remove the file in the <mount folder path>_cachedir/filesystem/<file path> (hradecFS will detect if the local cache file doesn't exist, and will retrieve it again automatically!)
2A. 1 folder - remove the control file in <mount folder path>_cachedir/control/<folder path control filename>.cacheDir (if the .cacheDir control file doesn't exist, hradecFS will retrieve the folder again from the remote filesystem)
2B. everything inside a folder - you have to remove the whole folder (including sub-files and sub-folders) int the <mount folder path>_cachedir/filesystem/<folder path>. In this case, hradecFS will known the .cacheDir control files are
wrong and will trigger a re-cache of the whole hierarchy bellow the folder.

Another way for folder hierarchy, is to delete all control files of all files and folders inside the one you want to refresh. If you want to just refresh the folders with new files/folders from the remote side, only delete .cacheDir
files below the folder hierarchy, including the one for the folder in question.